requirements.txt
zhipuai
python-dotenv
jupyter
openai==0.28
langchain[docarray]
pypdf
pandas
tiktoken
pydantic==1.10.9